backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
pydantic-settings==2.5.2
PyJWT[crypto]==2.9.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.12
openai>=1.60.0
httpx>=0.27.0,<1.0
google-generativeai==0.8.3
transformers==4.45.2
peft>=0.13.0
datasets>=3.0.0
rouge-score>=0.1.2
accelerate>=1.0.0
